msm: kgsl: Use GPU to switch pagetables when current one is default
Use the GPU to switch pagetables even if the current pagetable is
the default pagetable. This requires that commands are submitted to
GPU that do not belong to any context since the current context is
a NULL one. This is better for performance because if GPU is not
used to switch pagetable then synchronization is required between the
GPU and CPU.

Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>