Use pink defconfig
Previously this kernel only used the default defconfig which when compiled as is would leave out certain features of the kernel such as frandom or CPU governors. If this is updated, new people who download this source do not have to modify the config before compiling. This saves time, and confusion for newbs :p